lib: change order of "Post in" message (#345)
Change order to be slightly more consistent with previous message,
and so that the URL of the PR does not have to be followed directly
by a `:`. Some terminals consider the `:` to be part of the URL,
so clicking on it ends up on an invalid github page. Rearranging the
message so the URL is bounded by whitespace should work with all
terminals, and is no less readable.